Add copy color.
Copy color is now working on the GUI. When you select an area and hit
copy color the GUI will now calculate the stats for the area and paste
the results into the file buffer. The results include very good thresholds
for color tracking. This feature makes figuring out the thresholds for
binary() easy.

The IDE width and font were updated to make the pasted text look good.