Started improving python argument argument pasrsing.
In this commit I fixed up about half of the image library method
argument parsing front ends. The fixed up methods now:

1. Handle both standard and keyword arguments. If the user forgets to
pass the argument as a keyword but passes it in the right position the
method will "do the right" thing. If the user passes an argument twice
the method will take the keyword version of the value. Note that python
requires users to pass positional arguments first followed by keywords.
The new methods fix al previous problems about argument parsing.

2. If a drawing method takes an array the method can handle that array
being passed as a list of numbers now. This makes the drawing methods
much easier for newbies to use.

Anyway, I've managed to get about half way through the code. Need to
pivot to other things now. But, I will get back to this later. This fix
will in general allow us to remove all the warnings in the API about
having to call methods with keywords.