Add ability to turn AGC off and setting the temperature range
This is the main feature necessary for folks to be able to actually use
the lepton 3.5 effectively. I.e. they should set a temperature range
they want and then process the image based on that to get useful
information. When AGC is on it's great for a demo but rather worthless
for detecting things.