Debug FLIR Lepton support for 1.0/2.0/2.5/3.5
* Reduced the complexity of the FLIR Lepton code
* Made sure everything is reset on script stop
* Added example scripts
* Tested example scripts with FLIR Lepton 1.0/2.0/2.5/3.5

I had to try a rather large amount of stuff to get this all working. The
FLIR Lepton devices all behave slightly differently leading to a lot of
frustration. Originally this driver was going to be a lot smarter and
allow changing AGC without resetting the device however this proved to
be impossible along with various other things.