boards/OPENMVPT: Use IR rather than color image for temperature stats.
Before this change, the mean temperature of a blob is calculated based on the color values of the image. When running on the Pure Thermal this has the effect of labeling every blob with the min temperature of the entire scene. This change fixes this bug by using the IR image instead of the color image to get the right value.