ports/stm32: Fix DMA transfer of small buffer sizes.
HAL_DCMI_Start_DMA() only sets up one node in the circular linked-list
when the transfer size is < 64KB. This causes a user setting transfer
error which locks up the DMA hardware.

The fix is to dynamically make the link list linear on smaller sizes
and circular on larger sizes.